完成示例；问题修复与改进；

+++ uthread.h
@@ -26,19 +26,31 @@ typedef enum uthread_state {
   CREATED,
   WAITING_SIGNAL, // waiting for a condition/monitor lock
   WAITING_IO,     // waiting for an IO operation
-  YIELDED,        // uthread_yield has been called
+  YIELDED,        // uthread_yield has just been called
   RUNNING,
   STOPPED,
   ABORTED
 } uthread_state;
 
-uthread_executor_t *uthread_exec_create(unsigned long max_thread_count);
+// create an executor
+uthread_executor_t *uthread_exec_create(unsigned int max_thread_count);
+
+// create an uthread under specific executor
 int uthread_create(uthread_executor_t *executor, void (*func)(void *),
                    void *func_arg);
+
+// causes uthreads under the executor to be executed
+// the call will be blocked until all the uthreads exit or abort
 void uthread_exec_join(uthread_executor_t *executor);
+
+// destory specific executor
 void uthread_exec_destroy(uthread_executor_t *executor);
 
+// causes the calling uthread to yield execution to another uthread that is
+// ready to run on the current executor
 void uthread_yield();
+
+// causes the calling uthread to exit
 void uthread_exit();
 
 #ifdef __cplusplus
